Move sequence embedding to batching function
