+++ jax_unirep/evotuning.py
@@ -17,15 +17,16 @@ from jax_unirep.losses import _neg_cross_entropy_loss
 from .layers import mLSTM1900, mLSTM1900_HiddenStates
 from .optimizers import adamW
 from .utils import (
-    aa_seq_to_int,
-    batch_sequences,
+    # aa_seq_to_int,
+    # batch_sequences,
     dump_params,
     get_batching_func,
-    load_embedding_1900,
+    # load_embedding_1900,
     load_params,
-    one_hots,
+    # one_hots,
     right_pad,
     validate_mLSTM1900_params,
+    length_batch_input_outputs,
 )
 
 from joblib import Parallel, delayed
@@ -106,126 +107,6 @@ def avg_loss(
             num_seqs += len(x)
 
     return sum_loss / num_seqs
-
-
-def evotuning_pairs(s: str) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Given a sequence, return input-output pairs for evotuning.
-
-    The goal of evotuning is to get the RNN to accurately predict
-    the next character in a sequence.
-    This convenience function exists to prep a single sequence
-    into its corresponding input-output tensor pairs.
-
-    Given a 1D sequence of length `k`,
-    it gets represented as a 2D array of shape (k, 10),
-    where 10 is the size of the embedding of each amino acid,
-    and k-1 ranges from the zeroth a.a. to the nth a.a.
-    This is the first element in the returned tuple.
-
-    Given the same 1D sequence,
-    the output is defined as a 2D array of shape (k-1, 25),
-    where 25 is number of indices available to us
-    in `aa_to_int`,
-    and k-1 corresponds to the first a.a. to the nth a.a.
-    This is the second element in the returned tuple.
-
-    ### Parameters
-
-    - `s`: The protein sequence to featurize.
-
-    ### Returns
-
-    Two 2D NumPy arrays,
-    the first corresponding to
-    the input to evotuning with shape (n_letters, 10),
-    and the second corresponding to
-    the output amino acid to predict with shape (n_letters, 25).
-    """
-    seq_int = aa_seq_to_int(s[:-1])
-    next_letters_int = aa_seq_to_int(s[1:])
-    embeddings = load_embedding_1900()
-    x = onp.stack([embeddings[i] for i in seq_int])
-    y = onp.stack([one_hots[i] for i in next_letters_int])
-    return x, y
-
-
-def input_output_pairs(
-    sequences: List[str], n_cores: int = 1
-) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Generate input-output tensor pairs for evo-tuning.
-
-    We check that lengths of sequences are identical,
-    as this is necessary to ensure stacking of tensors happens correctly.
-
-    :param sequences: A list of sequences
-        to generate input-output tensor pairs.
-    :returns: Two NumPy arrays,
-        the first corresponding to the input to evotuning
-        with shape (n_sequences, n_letters+1, 10),
-        and the second corresponding to the output amino acids to predict
-        with shape (n_sequences, n_letters+1, 25).
-        Both will have an additional "sample" dimension as the first dim.
-    """
-    seqlengths = set(map(len, sequences))
-    logging.debug(seqlengths)
-    if not len(seqlengths) == 1:
-        raise ValueError(
-            """
-Sequences should be of uniform length, but are not.
-Please ensure that they are all of the same length before passing them in.
-"""
-        )
-
-    def embed_chunk(seq_chunk):
-        xs = []
-        ys = []
-
-        for s in seq_chunk:
-            x, y = evotuning_pairs(s)
-            xs.append(x)
-            ys.append(y)
-        return onp.stack(xs), onp.stack(ys)
-
-    chunk_size = int(len(sequences) / n_cores)
-    chunks = [
-        sequences[x : x + chunk_size]
-        for x in range(0, len(sequences), chunk_size)
-    ]
-
-    result = Parallel(n_jobs=n_cores, verbose=1)(
-        map(delayed(embed_chunk), chunks)
-    )
-    xs = onp.concatenate([r[0] for r in result])
-    ys = onp.concatenate([r[1] for r in result])
-    return xs, ys
-
-
-def length_batch_input_outputs(
-    sequences: Iterable[str],
-) -> Tuple[List[np.ndarray], List[np.ndarray]]:
-    """
-    Return lists of x and y tensors for evotuning, batched by their length.
-
-    This function exists because we need a way of
-    batching sequences by size conveniently.
-
-    :param sequences: A list of sequences to evotune on.
-    :returns: Two lists of NumPy arrays, one for xs and the other for ys.
-    """
-    idxs_batched = batch_sequences(sequences)
-
-    xs = []
-    ys = []
-    lens = []
-    for idxs in tqdm(idxs_batched):
-        seqs = [sequences[i] for i in idxs]
-        x, y = input_output_pairs(seqs)
-        xs.append(x)
-        ys.append(y)
-        lens.append(len(seqs[0]))
-    return xs, ys, lens
 
 
 def fit(
@@ -374,22 +255,24 @@ def fit(
             holdout_seqs = right_pad(holdout_seqs, max_len)
 
     # batch sequences by length
-    xs, ys, seq_lens = length_batch_input_outputs(sequences)
+    seqs_batched, seq_lens = length_batch_input_outputs(sequences)
     len_batching_funcs = {
-        sl: get_batching_func(x, y, batch_size)
-        for (sl, x, y) in zip(seq_lens, xs, ys)
+        sl: get_batching_func(seq_batch, batch_size)
+        for (sl, seq_batch) in zip(seq_lens, seqs_batched)
     }
 
     if holdout_seqs is not None:
-        holdout_xs, holdout_ys, holdout_seq_lens = length_batch_input_outputs(
+        holdout_seqs_batched, holdout_seq_lens = length_batch_input_outputs(
             holdout_seqs
         )
         holdout_len_batching_funcs = {
-            sl: get_batching_func(x, y, batch_size)
-            for (sl, x, y) in zip(holdout_seq_lens, holdout_xs, holdout_ys)
+            sl: get_batching_func(holdout_seq_batch, batch_size)
+            for (sl, holdout_seq_batch) in zip(
+                holdout_seq_lens, holdout_seqs_batched
+            )
         }
 
-    batch_lens = [len(batch) for batch in xs]
+    batch_lens = [len(batch) for batch in seqs_batched]
     logger.info(
         f"Length-batching done: "
         f"{len(batch_lens)} unique lengths, "
